requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.7
schedule==1.2.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pytz==2023.3
selenium==4.15.2
webdriver-manager==4.0.1
feedparser==6.0.10
tweepy==4.14.0 